Combine tasks into one

+++ template/.tools/scripts/merge_gitignore.ps1
@@ -1,9 +1,11 @@
-$original = Get-Content -Delimiter \0 -Path .gitignore
-$replace = Get-Content -Delimiter \0 -Path .tools/scripts/gitignore_to_merge
+if(Test-Path -Path .gitignore){
+    $original = Get-Content -Delimiter \0 -Path .gitignore
+    $replace = Get-Content -Delimiter \0 -Path .tools/scripts/gitignore_to_merge
 
-$lookbehind = '(?<=\s#\[\[\s)'  # "#[[\n" (have to use \s instead of \n)
-$match = '[\s\S]*'  # anything in-between
-$lookahead = '(?=\s#\]\]\s)'  # "#]]\n" (have to use \s instead of \n)
+    $lookbehind = '(?<=\s#\[\[\s)'  # "#[[\n" (have to use \s instead of \n)
+    $match = '[\s\S]*'  # anything in-between
+    $lookahead = '(?=\s#\]\]\s)'  # "#]]\n" (have to use \s instead of \n)
 
-$original -replace "$lookbehind$match$lookahead", $replace |
-    Set-Content -NoNewline -Path .gitignore
+    $original -replace "$lookbehind$match$lookahead", $replace |
+        Set-Content -NoNewline -Path .gitignore
+}
